Discovered failure mode
Outputs of just 0.5 for each maximizes entropy (naturally, given we're interpreting these as probabilities).

However an idea for how to proceed: when computing outputs, we somehow normalize the outputs based on their averages. Then you'd want to vary as widely as possible so the resulting probabilities would be near 0.5, and maximize your entropy. Not sure how to formalize that though. I'm imagining the expression binent(abs(y_i - E[y_i])) but I'm not sure quite what that means. Maybe something related to the "surprise" of using the encoding for E[y_i] and then actually recieving a coin with probability y_i and having to encode that?