Coding inefficiency for two strategies setup.
Illuminating results! It's optimal for each to condition on the signal, but if they do, then it doesn't matter if they are "anti-coordinating", because they're both conditioning on the same signal and so the conditional probability accounts for that. So the only interesting thing would be if there's multiple signals/sources of variation that they would anti-coordinate.