(In Progress) learned that my inefficiencyPayoffs is giving negative values...
I'm doing L(a_k, <a>) - H(<a>) and I'm wondering if previously I've actually done 

L(a_k, <a>) - H(a_k)

Both are coding inefficiencies. The second one says 

"the inefficiency of coding a_k as <a> not as a_k", which is kinda "the inefficiency of coding a_k, knowing just that it's an 'a'"

The first one is...

how many more bits you'd need if I gave you a_k and you had a coding scheme for <a> instead? But the fact that it's negative seems against the ethos. So now going to go back to commit with successfull evolution and see which one I used.